requirements.txt
marshmallow>=3
flask
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
inflection

apispec[yaml]
apispec-webframeworks
tox

# sentences
nltk
sentence-transformers

# needs to be removed
flask_cors
